src/backend/requirements-backend.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
shap>=0.44
joblib>=1.3
openpyxl>=3.1
matplotlib>=3.7
openai>=1.0
python-dotenv>=1.0
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9
psycopg[binary]>=3.2
